broke out goalinput and goalitem components
