+++ App.js
@@ -8,40 +8,24 @@ import {
   ScrollView,
   FlatList,
 } from "react-native";
+import GoalInput from "./components/GoalInput";
+import GoalItem from "./components/GoalItem";
 
 export default function App() {
-  const [enteredGoalText, setGoalText] = useState("");
   const [goalList, setGoalList] = useState([]);
 
-  function goalInputHandler(enteredText) {
-    setGoalText(enteredText);
-  }
-
-  function addGoalHandler() {
+  function addGoalHandler(enteredGoalText) {
     setGoalList((prevState) => [...prevState, { text: enteredGoalText }]);
-    setGoalText("");
   }
 
   return (
     <View style={styles.AppContainer}>
-      <View style={styles.InputContainer}>
-        <TextInput
-          style={styles.Input}
-          placeholder="Type goal here"
-          onChangeText={goalInputHandler}
-          value={enteredGoalText}
-        ></TextInput>
-        <Button title="Add Goal" onPress={addGoalHandler} />
-      </View>
+      <GoalInput onAddGoalHanlder={addGoalHandler} />
       <View style={styles.GoalsContainer}>
         <FlatList
           data={goalList}
           renderItem={(obj) => {
-            return (
-              <View style={styles.GoalItem}>
-                <Text style={styles.GoalItemText}>{obj.item.text}</Text>
-              </View>
-            );
+            return <GoalItem text={obj.item.text} />;
           }}
           keyExtractor={(item, index) => {
             return index;
@@ -57,34 +41,7 @@ const styles = StyleSheet.create({
     flex: 1,
     backgroundColor: "#fff",
   },
-  InputContainer: {
-    flex: 1,
-    flexDirection: "row",
-    alignItems: "center",
-    margin: 10,
-    padding: 20,
-    borderBottomWidth: 1,
-    borderBottomColor: "black",
-  },
-  Input: {
-    borderWidth: 1,
-    borderColor: "black",
-    padding: 5,
-    width: "70%",
-  },
   GoalsContainer: {
     flex: 6,
-  },
-  GoalItem: {
-    alignItems: "center",
-    borderRadius: 5,
-    borderWidth: 1,
-    borderColor: "#fff",
-    backgroundColor: "#4da0ff",
-    margin: 10,
-    padding: 10,
-  },
-  GoalItemText: {
-    color: "#fff",
   },
 });
